+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   coverage of the Platforms #977
 - Execute scripts to generate the documentation and standardize expid on documentation #1160
 
-### 4.1.15: Unreleased
+### 4.1.15: Bug fixes, enhancements, and new features
 
 The filter `-fp` of the command `autosubmit stats` changed in this release.
 Previously, a `-fp 0` would not raise any errors, and would bring all the
@@ -23,6 +23,8 @@ the filter the jobs. Not using any value for `-fp` still returns all jobs.
 - Deleted command `autosubmit test` that was not working in Autosubmit 4 #2386
 - Removed PBS and SGE platforms as they are not working in AS4 #2349
 - Log levels in the command line now accept `ERROR` #2412
+- Fix setstatus command to work in all cases #2381
+- Fix PS platform to work with the local machine #2374
 
 **Enhancements:**
 
@@ -30,6 +32,7 @@ the filter the jobs. Not using any value for `-fp` still returns all jobs.
 - Update portalocker requirement from <=3.1.1 to <=3.2.0 #2423
 - Fixed a `ZeroDivisionError` when using RO-Crate or `stats`, and also an issue
   where the message said `None` could not be iterable. #2389
+- Additional files are now generated upon using the `autosubmit inspect` command #2323'
 
 ### 4.1.14: Bug fixes, enhancements, and new features